requirements.txt
# Core
streamlit>=1.32.0
langchain-ollama
langchain-core
langgraph>=0.1.0
langgraph-checkpoint-sqlite
numpy>=1.24.0
requests

# RAG
faiss-gpu-cu12
sentence-transformers>=2.2.2

# File attachment support
pdfplumber
python-docx
openpyxl

# Evaluation
python-dotenv>=1.0.0
openai>=1.12.0
langchain-openai>=0.1.0
ragas>=0.2.0
datasets>=2.18.0
nest_asyncio>=1.6.0
